+++ codegens/libcurl/lib/util.js
@@ -160,7 +160,7 @@ const self = module.exports = {
   },
 
   /**
-   * Encode param except the following characters- [,{,},],%
+   * Encode param except the following characters- [,{,},],%,+
    *
    * @param {String} param
    * @returns {String}
@@ -171,6 +171,7 @@ const self = module.exports = {
       .replace(/%7B/g, '{')
       .replace(/%5D/g, ']')
       .replace(/%7D/g, '}')
+      .replace(/%2B/g, '+')
       .replace(/%25/g, '%')
       .replace(/'/g, '%27');
   },