+++ codegens/libcurl/test/unit/convert.test.js
@@ -2,6 +2,7 @@ var expect = require('chai').expect,
   sdk = require('postman-collection'),
   convert = require('../../index').convert,
   getOptions = require('../../index').getOptions,
+  getUrlStringfromUrlObject = require('../../lib/util').getUrlStringfromUrlObject,
   sanitize = require('../../lib/util').sanitize,
   mainCollection = require('../../../../test/codegen/newman/fixtures/basicCollection.json');
 
@@ -145,6 +146,38 @@ describe('libcurl convert function', function () {
         expect(snippet).to.include('curl_mime_name(part, "invalid src");');
       });
     });
+
+    it('should free up headers list after request is sent', function () {
+      var request = new sdk.Request({
+        'method': 'GET',
+        'header': [
+          {
+            'key': 'Accept',
+            'value': 'application/json'
+          },
+          {
+            'key': 'Content-Type',
+            'value': 'application/json'
+          }
+        ],
+        'url': {
+          'raw': 'https://google.com',
+          'protocol': 'https',
+          'host': [
+            'google',
+            'com'
+          ]
+        }
+      });
+      convert(request, {}, function (error, snippet) {
+        if (error) {
+          expect.fail(null, null, error);
+        }
+        expect(snippet).to.be.a('string');
+        expect(snippet).to.include('curl_slist_free_all(headers)');
+      });
+    });
+
   });
 
   describe('getOptions function', function () {
@@ -175,5 +208,14 @@ describe('libcurl convert function', function () {
       expect(sanitize('inputString     ', true)).to.equal('inputString');
     });
 
+    it('should not encode unresolved query params and ' +
+    'encode every other query param, both present together', function () {
+      let rawUrl = 'https://postman-echo.com/get?key1={{value}}&key2=\'a b+c\'',
+        urlObject = new sdk.Url(rawUrl),
+        outputUrlString = getUrlStringfromUrlObject(urlObject);
+      expect(outputUrlString).to.not.include('key1=%7B%7Bvalue%7B%7B');
+      expect(outputUrlString).to.not.include('key2=\'a b+c\'');
+      expect(outputUrlString).to.equal('https://postman-echo.com/get?key1={{value}}&key2=%27a%20b+c%27');
+    });
   });
 });